[feature] #19 bean dependency cycle 해결 및 토큰 유효 검사 코드 추가
